app/requirements.txt
flask==2.2.2
Scrapy==2.9.0
psycopg2-binary==2.9.6